Refactor illustration step into smaller components
