+++ src/components/book-customization/illustrations/options.ts
@@ -1,0 +1,30 @@
+export const skinTones = [
+  { value: "light", label: "Light", color: "#FFE0BD" },
+  { value: "medium-light", label: "Medium Light", color: "#F1C27D" },
+  { value: "medium", label: "Medium", color: "#E0AC69" },
+  { value: "medium-dark", label: "Medium Dark", color: "#C68642" },
+  { value: "dark", label: "Dark", color: "#8D5524" },
+];
+
+export const hairColors = [
+  { value: "black", label: "Black", color: "#252525" },
+  { value: "brown", label: "Brown", color: "#6A4E42" },
+  { value: "blonde", label: "Blonde", color: "#FFF5E1" },
+  { value: "red", label: "Red", color: "#D36B56" },
+  { value: "gray", label: "Gray", color: "#AAAAAA" },
+];
+
+export const hairStyles = [
+  { value: "short", label: "Short" },
+  { value: "medium", label: "Medium" },
+  { value: "long", label: "Long" },
+  { value: "curly", label: "Curly" },
+  { value: "wavy", label: "Wavy" },
+];
+
+export const backgrounds = [
+  { value: "home", label: "Home", image: "https://images.unsplash.com/photo-1560448204-e02f11c3d0e2?auto=format&fit=crop&q=80&w=2670&ixlib=rb-4.0.3" },
+  { value: "garden", label: "Garden", image: "https://images.unsplash.com/photo-1560429502-d55f5eb98848?auto=format&fit=crop&q=60&w=800&ixlib=rb-4.0.3" },
+  { value: "beach", label: "Beach", image: "https://images.unsplash.com/photo-1507525428034-b723cf961d3e?auto=format&fit=crop&q=80&w=2673&ixlib=rb-4.0.3" },
+  { value: "park", label: "Park", image: "https://images.unsplash.com/photo-1519331379826-f10be5486c6f?auto=format&fit=crop&q=80&w=2672&ixlib=rb-4.0.3" },
+];
